minor changes to menu function
Changed the max number of test cases allowed to be 100 and numbers per
test case to 200.  I also went back to passing out the max_value and
min_value for the range of the numbers randomly generated to account for
the possibility of a range beginning at a negative number.